amélioré causal dispatcher et adapter a ce que stéphane à modifié
git-svn-id: https://scm.gforge.inria.fr/svn/rtcelog/software/JBenchmarkerCE@1396 47090b3f-bbb5-42d7-8af4-003170c53eba

+++ src/main/java/jbenchmarker/MainSimulaReplica.java
@@ -120,7 +120,7 @@ public class MainSimulaReplica {
         writeFile(calculAverag(rtime), "gen", nameUsr);
         writeFile(calculAverag(mem), "mem", nameUsr);
     }
-    private static void toArrayDouble(double[] tab, List<Long> memUsed) {
+    private static void toArrayDouble(double[] tab, List<Double> memUsed) {
         for(int i=0; i<memUsed.size(); i++)
             tab[i] = memUsed.get(i).doubleValue();
     }
